added separate ThemeConfiguration
